Refactore 1.02
